requirements.txt
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
playwright==1.41.2
transformers==4.37.2
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
scikit-learn==1.4.0
pyyaml==6.0.1
emoji==2.10.1
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
